+++ tool/ios/build_installed_sdk.sh
@@ -34,7 +34,7 @@ export TARGET_LIB=$target_lib
 
 sh "$script_directory/build-runtime-closure.sh" iphoneos
 
-framework_source_sha256='c20edc77779c24c411854a19d234887615a6ba0a352784d35a970fb0a7d148a5'
+framework_source_sha256='253c523dc0fb2d9ea931e82f3b9174613ff6519e3718ad650f550ae02593d4dc'
 framework_archive_source="$script_directory/bonsai_flutter.tar.gz"
 framework_archive="$work_root/bonsai_flutter.tar.gz"
 framework_source="$work_root/framework-source"